+++ src/main/java/io/fabric8/maven/docker/AbstractDockerMojo.java
@@ -1,10 +1,12 @@
 package io.fabric8.maven.docker;
 
 import java.io.File;
-import java.util.*;
+import java.io.IOException;
+import java.util.Date;
+import java.util.List;
+import java.util.Properties;
 
 import io.fabric8.maven.docker.access.DockerAccess;
-import io.fabric8.maven.docker.access.DockerAccessException;
 import io.fabric8.maven.docker.access.ExecException;
 import io.fabric8.maven.docker.config.BuildImageConfiguration;
 import io.fabric8.maven.docker.config.ConfigHelper;
@@ -23,9 +25,9 @@ import io.fabric8.maven.docker.service.ServiceHubFactory;
 import io.fabric8.maven.docker.util.AnsiLogger;
 import io.fabric8.maven.docker.util.AuthConfigFactory;
 import io.fabric8.maven.docker.util.EnvUtil;
+import io.fabric8.maven.docker.util.GavLabel;
 import io.fabric8.maven.docker.util.ImageNameFormatter;
 import io.fabric8.maven.docker.util.Logger;
-import io.fabric8.maven.docker.util.PomLabel;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecution;
@@ -211,17 +213,17 @@ public abstract class AbstractDockerMojo extends AbstractMojo implements Context
 
             ConfigHelper.validateExternalPropertyActivation(project, images);
 
-            // The 'real' images configuration to use (configured images + externally resolved images)
-            this.minimalApiVersion = initImageConfiguration(getBuildTimestamp());
             DockerAccess access = null;
             try {
+                // The 'real' images configuration to use (configured images + externally resolved images)
+                this.minimalApiVersion = initImageConfiguration(getBuildTimestamp());
                 if (isDockerAccessRequired()) {
                     DockerAccessFactory.DockerAccessContext dockerAccessContext = getDockerAccessContext();
                     access = dockerAccessFactory.createDockerAccess(dockerAccessContext);
                 }
                 ServiceHub serviceHub = serviceHubFactory.createServiceHub(project, session, access, log, logSpecFactory);
                 executeInternal(serviceHub);
-            } catch (DockerAccessException | ExecException exp) {
+            } catch (IOException | ExecException exp) {
                 logException(exp);
                 throw new MojoExecutionException(log.errorMessage(exp.getMessage()), exp);
             } catch (MojoExecutionException exp) {
@@ -271,7 +273,7 @@ public abstract class AbstractDockerMojo extends AbstractMojo implements Context
      * call or a new current date is created
      * @return timestamp to use
      */
-    protected synchronized Date getBuildTimestamp() throws MojoExecutionException {
+    protected synchronized Date getBuildTimestamp() throws IOException {
         Date now = (Date) getPluginContext().get(CONTEXT_KEY_BUILD_TIMESTAMP);
         if (now == null) {
             now = getReferenceDate();
@@ -282,7 +284,7 @@ public abstract class AbstractDockerMojo extends AbstractMojo implements Context
 
     // Get the reference date for the build. By default this is picked up
     // from an existing build date file. If this does not exist, the current date is used.
-    protected Date getReferenceDate() throws MojoExecutionException {
+    protected Date getReferenceDate() throws IOException {
         Date referenceDate = EnvUtil.loadTimestamp(getBuildTimestampFile());
         return referenceDate != null ? referenceDate : new Date();
     }
@@ -352,7 +354,7 @@ public abstract class AbstractDockerMojo extends AbstractMojo implements Context
      * @param serviceHub context for accessing backends
      */
     protected abstract void executeInternal(ServiceHub serviceHub)
-        throws DockerAccessException, ExecException, MojoExecutionException;
+        throws IOException, ExecException, MojoExecutionException;
 
     // =============================================================================================
 
@@ -384,9 +386,9 @@ public abstract class AbstractDockerMojo extends AbstractMojo implements Context
 
     // =================================================================================
 
-    protected PomLabel getPomLabel() {
+    protected GavLabel getGavLabel() {
         // Label used for this run
-        return new PomLabel(project.getGroupId(),project.getArtifactId(),project.getVersion());
+        return new GavLabel(project.getGroupId(), project.getArtifactId(), project.getVersion());
     }
 
     protected LogDispatcher getLogDispatcher(ServiceHub hub) {