+++ src/main/java/io/fabric8/maven/docker/service/BuildService.java
@@ -287,7 +287,7 @@ public class BuildService {
 
         public static class Builder {
 
-            private BuildContext context = new BuildContext();
+            private BuildContext context;
 
             public Builder() {
                 this.context = new BuildContext();